requirements.txt
flask==1.0.2
Flask-Cors==3.0.4
pytest==3.6.0
pytest-flask==0.10.0
PyYAML==3.13
